chore: code clean and minor fixes
